final/requirements.txt
ultralytics
opencv-python
torch
torchvision
pillow
numpy
scikit-learn
